Move interfaces out into their own package
