+++ src/main/java/sim/view/KitchenGUI.java
@@ -5,6 +5,7 @@ import javax.swing.JTextArea;
 import javax.swing.border.Border;
 
 import sim.app.Order;
+import sim.interfaces.Observer;
 import sim.model.Kitchen;
 
 public class KitchenGUI extends JTextArea implements Observer {
