+++ src/main/java/sim/view/ServerGUI.java
@@ -21,6 +21,7 @@ import javax.swing.event.ChangeEvent;
 import javax.swing.event.ChangeListener;
 
 import sim.app.Order;
+import sim.interfaces.Observer;
 import sim.model.Server;
 
 public class ServerGUI extends JPanel implements Observer {
@@ -29,15 +30,15 @@ public class ServerGUI extends JPanel implements Observer {
 
 	private Server server;
 	private int number;
-	
+
 	private JTextArea serverArea;
 	private JSlider serverSlider;
 	private JLabel label;
-	
+
 	static final int MIN = -10;
 	static final int MAX = 10;
 	static final int INIT = 0;
-	
+
 	private Long servingTime;
 
 	public ServerGUI(Server server) {
@@ -46,13 +47,13 @@ public class ServerGUI extends JPanel implements Observer {
 		server.registerObserver(this);
 		setLayout(new GridLayout(2,1));
 		add(setupControls());
-		
+
 
 		setup();
 	}
 
 	public void setup() {
-		
+
 		serverArea = new JTextArea();
 		serverArea.setEditable(false);
 		Border border1 = BorderFactory.createTitledBorder("Server " + number);
@@ -60,13 +61,13 @@ public class ServerGUI extends JPanel implements Observer {
 		serverPane.setBorder(border1);
 		add(serverPane);
 	}
-	
-	
+
+
 	public JPanel setupControls() {
 		JPanel controls = new JPanel();
 		Border margin = new EmptyBorder(10,0,0,0);
 		controls.setLayout(new BoxLayout(controls, BoxLayout.Y_AXIS));
-		
+
 		// displays serving speed of a server
 		label = new JLabel();
 		label.setFont(new Font(null, Font.BOLD, 13));
@@ -76,33 +77,33 @@ public class ServerGUI extends JPanel implements Observer {
 		servingTime = server.getBaseSpeed()/1000;
 		String text  = "Serving Speed: " + servingTime + "s";
 		label.setText(text);
-		
-		// add slider and slider label to a JPanel 
+
+		// add slider and slider label to a JPanel
 		controls.add(label);
 		controls.add(setupSlider());
-		
+
 		return controls;
 	}
-	
+
 	// sets up server speed controls (i.e., the slider)
 	public JSlider setupSlider() {
 		serverSlider = new JSlider(JSlider.HORIZONTAL, MIN, MAX, INIT);
-		
+
 		// slider formatting
 		serverSlider.setMajorTickSpacing(5);
 		serverSlider.setPaintTicks(true);
 		serverSlider.setPaintLabels(true);
-		
+
 		// create a table with custom slider labels
 		Hashtable<Integer, JLabel> labelTable = new Hashtable<Integer, JLabel>();
 		labelTable.put( MIN, new JLabel("Slow") );
 		labelTable.put( INIT, new JLabel("Normal") );
 		labelTable.put( MAX, new JLabel("Fast") );
 		serverSlider.setLabelTable( labelTable );
-		
+
 		return serverSlider;
 	}
-	
+
 	// add listener to update button
 	public void addSetListener(ChangeListener ce) {
 		serverSlider.addChangeListener(ce);
